fix(run-pytest): Make the code and the .coverage file accessible in the current workdir
Github actions does not works well when WORKDIR is defined within a Docker image
which is our case most of the time.
To counter this we override the "working-directory" input from script steps (steps that 'run' a shell script)
but this does not work at all for steps that use custom action like dorny/test-reporter
or py-cov-action/python-coverage-comment-action .

dorny/test-reporter offers a workaround by exposing a working-directory input too, but
py-cov-action/python-coverage-comment-action does not and the code is made so that
any of its path input MUST be within the current working directory: this means that this step does
dot see our code (which leaves in inputs.workdir) and so does not see the .coverage file either.

To work around it, we need to checkout the code in the current working directory forced by github action
when it creates the container (something like /github/workspace) and we copy the .coverage file
generated during the test step into this working directory.

I tried to implement this workaround directly within the py-cov-action/python-coverage-comment-action project
but as it also uses docker to run itself, things go crazy and I never manage to make something working
without refactoring the all codebase.

See:
https://docs.github.com/en/actions/sharing-automations/creating-actions/dockerfile-support-for-github-actions#workdir